requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
# torch>=2.0.0  # Descomentar si se necesita PyTorch

# Data processing
scipy>=1.11.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.0
